+++ connect/mirror/src/test/java/org/apache/kafka/connect/mirror/MirrorCheckpointConnectorTest.java
@@ -17,7 +17,6 @@
 package org.apache.kafka.connect.mirror;
 
 import org.apache.kafka.clients.admin.ConsumerGroupListing;
-import org.apache.kafka.common.config.ConfigDef;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -32,8 +31,6 @@ import java.util.stream.Collectors;
 
 import static org.apache.kafka.connect.mirror.TestUtils.makeProps;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.spy;
@@ -42,57 +39,6 @@ import static org.mockito.Mockito.spy;
 public class MirrorCheckpointConnectorTest {
 
     private static final String CONSUMER_GROUP = "consumer-group-1";
-
-    @Test
-    public void testTaskConfigConsumerGroups() {
-        List<String> groups = Arrays.asList("consumer-1", "consumer-2", "consumer-3");
-        MirrorCheckpointConfig config = new MirrorCheckpointConfig(makeProps());
-        Map<String, String> props = config.taskConfigForConsumerGroups(groups);
-        MirrorCheckpointTaskConfig taskConfig = new MirrorCheckpointTaskConfig(props);
-        assertEquals(taskConfig.taskConsumerGroups(), new HashSet<>(groups),
-                "Setting consumer groups property configuration failed");
-    }
-
-    @Test
-    public void testGroupMatching() {
-        MirrorCheckpointConfig config = new MirrorCheckpointConfig(makeProps("groups", "group1"));
-        assertTrue(config.groupFilter().shouldReplicateGroup("group1"),
-                "topic1 group matching property configuration failed");
-        assertFalse(config.groupFilter().shouldReplicateGroup("group2"),
-                "topic2 group matching property configuration failed");
-    }
-
-    @Test
-    public void testNonMutationOfConfigDef() {
-        // Sanity check to make sure that these properties are actually defined for the task config,
-        // and that the task config class has been loaded and statically initialized by the JVM
-        ConfigDef taskConfigDef = MirrorCheckpointTaskConfig.TASK_CONFIG_DEF;
-        assertTrue(
-                taskConfigDef.names().contains(MirrorCheckpointConfig.TASK_CONSUMER_GROUPS),
-                MirrorCheckpointConfig.TASK_CONSUMER_GROUPS + " should be defined for task ConfigDef"
-        );
-
-        // Ensure that the task config class hasn't accidentally modified the connector config
-        assertFalse(
-                MirrorCheckpointConfig.CONNECTOR_CONFIG_DEF.names().contains(MirrorCheckpointConfig.TASK_CONSUMER_GROUPS),
-                MirrorCheckpointConfig.TASK_CONSUMER_GROUPS + " should not be defined for connector ConfigDef"
-        );
-    }
-
-    @Test
-    public void testConsumerConfigsForOffsetSyncsTopic() {
-        Map<String, String> connectorProps = makeProps(
-                "source.consumer.max.partition.fetch.bytes", "1",
-                "target.consumer.heartbeat.interval.ms", "1",
-                "consumer.max.poll.interval.ms", "1",
-                "fetch.min.bytes", "1"
-        );
-        MirrorCheckpointConfig config = new MirrorCheckpointConfig(connectorProps);
-        assertEquals(config.sourceConsumerConfig(), config.offsetSyncsTopicConsumerConfig());
-        connectorProps.put("offset-syncs.topic.location", "target");
-        config = new MirrorCheckpointConfig(connectorProps);
-        assertEquals(config.targetConsumerConfig(), config.offsetSyncsTopicConsumerConfig());
-    }
 
     @Test
     public void testMirrorCheckpointConnectorDisabled() {
